backend/app/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
pytesseract
PyMuPDF
pydantic
